safari-600.7-branch/LayoutTests/sputnik/Conformance/11_Expressions/11.3_PostfixExpressions/11.3.2_Postfix_Decrement_Operator/S11.3.2_A2.1_T3-expected.txt

S11.3.2_A2.1_T3

PASS ReferenceError: Postfix -- operator applied to value that is not a reference.

TEST COMPLETE
